backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
groq==0.4.1
gTTS==2.5.0
moviepy==1.0.3
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1
